+++ admin/partials/lkn-antispam-for-givewp-misc-functions.php
@@ -11,8 +11,7 @@ if ( ! defined('WPINC')) {
  *
  * @since
  */
-function verifyPluginDependencies(): void
-{
+function verifyPluginDependencies(): void {
     // Load plugin helper functions.
     if ( ! function_exists('deactivate_plugins') || ! function_exists('is_plugin_active')) {
         require_once ABSPATH . '/wp-admin/includes/plugin.php';
@@ -21,7 +20,7 @@ function verifyPluginDependencies(): void
     // Flag to check whether deactivate plugin or not.
     $is_deactivate_plugin = null;
 
-    $give_antispam_path = ABSPATH . '/wp-content/plugins/give-antispam/give-antispam.php';
+    $lkn_antispam_path = ABSPATH . '/wp-content/plugins/give-antispam/lkn-antispam-for-givewp.php';
 
     $is_installed = false;
 
@@ -54,7 +53,7 @@ function verifyPluginDependencies(): void
 
     // Deactivate plugin.
     if ($is_deactivate_plugin) {
-        deactivate_plugins($give_antispam_path);
+        deactivate_plugins($lkn_antispam_path);
 
         if (isset($_GET['activate'])) {
             unset($_GET['activate']);
@@ -62,8 +61,7 @@ function verifyPluginDependencies(): void
     }
 }
 
-function __give_lkn_antispam_dependency_notice(): void
-{
+function __lkn_antispam_for_givewp_dependency_notice(): void {
     // Admin notice.
     $message = sprintf(
         '<div class="notice notice-error"><p><strong>%1$s</strong> %2$s <a href="%3$s" target="_blank">%4$s</a>  %5$s %6$s+ %7$s.</p></div>',
@@ -79,9 +77,8 @@ function __give_lkn_antispam_dependency_notice(): void
     echo $message;
 }
 
-function dependencyAlert(): void
-{
-    add_action('admin_notices', '__give_lkn_antispam_dependency_notice');
+function dependencyAlert(): void {
+    add_action('admin_notices', '__lkn_antispam_for_givewp_dependency_notice');
 }
 
 /**
@@ -89,8 +86,7 @@ function dependencyAlert(): void
  *
  * @since
  */
-function __give_lkn_antispam_inactive_notice(): void
-{
+function __lkn_antispam_for_givewp_inactive_notice(): void {
     // Admin notice.
     $message = sprintf(
         '<div class="notice notice-error"><p><strong>%1$s</strong> %2$s <a href="%3$s" target="_blank">%4$s</a> %5$s.</p></div>',
@@ -104,9 +100,8 @@ function __give_lkn_antispam_inactive_notice(): void
     echo $message;
 }
 
-function inactiveAlert(): void
-{
-    add_action('admin_notices', '__give_lkn_antispam_inactive_notice');
+function inactiveAlert(): void {
+    add_action('admin_notices', '__lkn_antispam_for_givewp_inactive_notice');
 }
 
 /**
@@ -119,8 +114,7 @@ function inactiveAlert(): void
  *
  * @return array
  */
-function __give_lkn_antispam_plugin_row_meta($plugin_meta)
-{
+function __lkn_antispam_for_givewp_plugin_row_meta($plugin_meta) {
     $new_meta_links['setting'] = sprintf(
         '<a href="%1$s">%2$s</a>',
         admin_url('edit.php?post_type=give_forms&page=give-settings&tab=general&section=access-control'),
@@ -129,28 +123,3 @@ function __give_lkn_antispam_plugin_row_meta($plugin_meta)
 
     return array_merge($plugin_meta, $new_meta_links);
 }
-
-// APAGAR DEPOIS
-// /**
-//  * Show activation banner.
-//  *
-//  * @since
-//  */
-// function __give_lkn_antispam_activation(): void
-// {
-//     // Initialize activation welcome banner.
-//     if (class_exists('Lkn_Give_Antispam')) {
-//         // Only runs on admin.
-//         $args = array(
-//             'file' => LKN_ANTISPAM_FOR_GIVEWP_FILE,
-//             'name' => __('Antispam', 'antispam-donation-for-givewp'),
-//             'version' => LKN_ANTISPAM_FOR_GIVEWP_VERSION,
-//             'settings_url' => admin_url('edit.php?post_type=give_forms&page=give-settings&tab=general&section=access-control'),
-//             'documentation_url' => 'https://www.linknacional.com.br/wordpress/givewp/',
-//             'support_url' => 'https://www.linknacional.com.br/wordpress/givewp/',
-//             'testing' => false, // Never leave true.
-//         );
-
-//         new Lkn_Give_Antispam($args);
-//     }
-// }